+++ src/components/forums/CreateForum.js
@@ -95,7 +95,7 @@ const CreateForum = () => {
     }
 
     const [forumTag, setForumTag] = React.useState('');
-    const forumTagList = ["School", "Co-op", "Funny", "Debate", "Rant", "Interview", "Class Review", "Good News"];
+    const forumTagList = ["","School", "Co-op", "Funny", "Debate", "Rant", "Interview", "Class Review", "Good News"];
 
     const handleForumTag = (event) => {
         setForumTag(event.target.value);
